pyproject.toml
[project]
name = "newsblur"
version = "0.1.0"
requires-python = ">=3.13,<3.14"
dependencies = [
    "amqp==5.3.1",
    "appdirs==1.4.4",
    "anthropic==0.75.0",
    "asgiref==3.11.1",
    "attrs==25.4.0",
    "beautifulsoup4==4.14.3",
    "billiard==4.2.4",
    "bleach==6.3.0",
    "boto3==1.42.46",
    "botocore==1.42.46",
    "black==26.1.0",
    "celery==5.6.2",
    "certifi==2026.1.4",
    "cffi==2.0.0",
    "chardet==5.2.0",
    "ConfigArgParse==1.7.1",
    "cryptography==46.0.5",
    "cssutils==2.11.1",
    "distlib==0.4.0",
    "Django>=4.2,<5.0",
    "django-anymail==14.0",
    "django-compress==1.0.1",
    "django-cors-headers==4.9.0",
    "django-extensions==4.1",
    "django-oauth-toolkit==3.2.0",
    "django-paypal==2.1",
    "django-qurl==0.1.1",
    "django-pipeline==4.1.0",
    "django-prometheus==2.4.1",
    "django-redis==6.0.0",
    "django-redis-sessions==0.6.2",
    "django-ses==4.7.1",
    "django-timezone-field==7.2.1",
    "dnspython==2.8.0",
    "dopy==0.3.7",
    "elasticsearch>=7.17.0,<8",
    "factory-boy==3.3.3",
    "Faker==40.4.0",
    "feedparser>=6,<7",
    "filelock==3.20.3",
    "flake8==7.3.0",
    "Flask==3.1.2",
    "Flask-BasicAuth==0.2.0",
    "gunicorn==25.0.3",
    "gevent==25.9.1",
    "greenlet==3.3.1",
    "google-api-python-client==2.190.0",
    "google-auth==2.48.0",
    "google-auth-httplib2==0.2.0",
    "hiredis==3.3.0",
    "httplib2==0.31.2",
    "idna==3.11",
    "image==1.5.33",
    "iniconfig==2.3.0",
    "isodate==0.7.2",
    "isort==7.0.0",
    "Jinja2==3.1.6",
    "jmespath==1.1.0",
    "jsonpickle==4.1.1",
    "kombu==5.6.2",
    "locust==2.43.2",
    "lxml==6.0.2",
    "lxml-html-clean==0.4.3",
    "mongoengine==0.29.1",
    "msgpack==1.1.2",
    "ndg-httpsclient==0.5.1",
    "numpy==2.4.2",
    "oauth2==1.9.0.post1",
    "oauthlib==3.3.1",
    "openai~=1.56.1",
    "google-genai==1.62.0",
    "lightningcss==0.3.1",
    "paypalrestsdk==1.13.3",
    "pbr==7.0.3",
    "Pillow==12.1.1",
    "pluggy==1.6.0",
    "psutil==7.2.2",
    "psycopg2==2.9.11",
    "pyapns-client==2.0.6",
    "pyasn1==0.6.2",
    "pycparser==3.0",
    "pymongo>=4.9,<4.14",
    "PyMySQL==1.1.2",
    "pynliner==0.8.0",
    "pyOpenSSL==25.3.0",
    "pyparsing==3.3.2",
    "PySocks==1.7.1",
    "pytest==9.0.2",
    "pytest-django==4.11.1",
    "pytest-cov==7.0.0",
    "python-dateutil==2.9.0.post0",
    "python-digitalocean==1.17.0",
    "python-gflags==3.1.2",
    "pytz==2025.2",
    "PyYAML==6.0.3",
    "pyzmq==27.1.0",
    "redis>=5.0.0,<6",
    "requests>=2.31.0,<3",
    "requests-oauthlib==2.0.0",
    "scipy==1.17.0",
    "sentry-sdk==2.52.0",
    "sgmllib3k==1.0.0",
    "simplejson==3.20.2",
    "smartypants==2.0.2",
    "soupsieve==2.8.3",
    "sqlparse==0.5.5",
    "stevedore==5.6.0",
    "stripe==14.3.0",
    "subdomains==3.0.1",
    "text-unidecode==1.3",
    "tiktoken~=0.4.0",
    "toml==0.10.2",
    "urllib3>=2.0.0,<3",
    "url-normalize>=1,<2",
    "vine==5.1.0",
    "virtualenv==20.36.1",
    "virtualenv-clone==0.5.7",
    "virtualenvwrapper==6.1.1",
    "webencodings==0.5.1",
    "XlsxWriter==3.2.9",
    "zope.event==6.1",
    "zope.interface==8.2",
]
